+++ ssnolib/dcat/resource.py
@@ -10,8 +10,9 @@ from datetime import datetime
 from typing import Union, List, Optional
 
 from dateutil import parser
-from ontolutils import Thing, as_id, urirefs, namespaces
-from ontolutils.classes.typing import BlankNodeType
+from ontolutils import Thing, as_id, urirefs, namespaces, LangString
+from ontolutils.typing import BlankNodeType
+from ontolutils.typing import ResourceType
 from pydantic import HttpUrl, FileUrl, field_validator, Field, model_validator
 
 from ssnolib.utils import download_file
@@ -52,46 +53,37 @@ class Resource(Thing):
         Version of the resource (dcat:version),
         best following semantic versioning (https://semver.org/lang/de/)
     """
-    title: str = None  # dcterms:title
-    description: str = None  # dcterms:description
+    title: Optional[Union[LangString, List[LangString]]] = None  # dcterms:title
+    description: Optional[Union[LangString, List[LangString]]] = None  # dcterms:description
     creator: Union[
-        foaf.Agent, foaf.Organization, foaf.Person, prov.Person, prov.Agent, prov.Organization, HttpUrl, BlankNodeType,
-        List[Union[
-            foaf.Agent, foaf.Organization, foaf.Person, prov.Person, prov.Agent, prov.Organization, HttpUrl, BlankNodeType]]
+        foaf.Agent,
+        foaf.Organization,
+        foaf.Person,
+        prov.Person,
+        prov.Agent,
+        prov.Organization,
+        HttpUrl,
+        BlankNodeType,
+        List[
+            Union[
+                foaf.Agent,
+                foaf.Organization,
+                foaf.Person,
+                prov.Person,
+                prov.Agent,
+                prov.Organization,
+                HttpUrl,
+                BlankNodeType
+            ]
+        ]
     ] = None  # dcterms:creator
     version: str = None  # dcat:version
-    identifier: Union[str, HttpUrl] = None  # dcterms:identifier
+    identifier: str = None  # dcterms:identifier
 
     @model_validator(mode="before")
     def change_id(self):
         """Change the id to the downloadURL"""
         return as_id(self, "identifier")
-
-    # @field_validator('creator', mode='before')
-    # @classmethod
-    # def _parse_creator(cls, creator):
-    #     # check if creator is a valid person or organization. if both fail, just pass creator data, it will fail later
-    #     is_person = False
-    #     is_organisation = False
-    #     try:
-    #         person = foaf.Person.model_validate(creator, strict=True)
-    #         is_person = True
-    #     except pydantic.ValidationError:
-    #         pass
-    #         # not a person
-    #     try:
-    #         organisation = foaf.Organization.model_validate(creator, strict=True)
-    #         is_organisation = True
-    #     except pydantic.ValidationError:
-    #         # not an organisation
-    #         pass
-    #     if is_person and is_organisation:
-    #         return creator  # cannot distinguish between person and organisation
-    #     if is_person:
-    #         return person
-    #     if is_organisation:
-    #         return organisation
-    #     return creator
 
     @field_validator('identifier', mode='before')
     @classmethod
@@ -120,7 +112,7 @@ class Distribution(Resource):
     ----------
     downloadURL: Union[HttpUrl, FileUrl]
         Download URL of the distribution (dcat:downloadURL)
-    mediaType: HttpUrl = None
+    mediaType: ResourceType = None
         Media type of the distribution (dcat:mediaType).
         Should be defined by the [IANA Media Types registry](https://www.iana.org/assignments/media-types/media-types.xhtml)
     byteSize: int = None
@@ -130,7 +122,7 @@ class Distribution(Resource):
     """
     downloadURL: Union[HttpUrl, FileUrl, pathlib.Path] = Field(default=None, alias='download_URL')
     accessURL: Union[HttpUrl, FileUrl, pathlib.Path] = Field(default=None, alias='access_URL')
-    mediaType: HttpUrl = Field(default=None, alias='media_type')  # dcat:mediaType
+    mediaType: ResourceType = Field(default=None, alias='media_type')  # dcat:mediaType
     byteSize: int = Field(default=None, alias='byte_size')  # dcat:byteSize
     keyword: List[str] = None  # dcat:keyword