+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Log of changes in the versions
 
 - update `to_xml`
 - dcat:Distribution allows foaf-Agents as creators
+- fix StandardNameTable version property. hasVersion was used wrongly as a string, now it is a resource and
+  `schema:version` is added for the actual version.
 
 ## v1.5.1.3